html/semantics/forms/textfieldselection/selection-after-content-change.html

<!doctype html>
<meta charset="utf-8">
<title>Selection indices after content change</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<input id="i1" type="text" value="hello">
<textarea id="t1">hello</textarea>

<script>
"use strict";

runTest("input out of document", () => {
  const input = document.createElement("input");
  input.value = "hello";
  return input;
});

runTest("input in document", () => {
  return document.querySelector("#i1");
});

runTest("input in document, with focus", () => {
  const input = document.querySelector("#i1");
  input.value = "hello";
  input.focus();
  return input;
});

runTest("textarea out of document", () => {
  const textarea = document.createElement("textarea");
  textarea.value = "hello";
  return textarea;
});

runTest("textarea in document", () => {
  return document.querySelector("#t1");
});

runTest("textarea in document, with focus", () => {
  const textarea = document.querySelector("#t1");
  textarea.value = "hello";
  textarea.focus();
  return textarea;
});

function runTest(descriptor, elementFactory) {
  test(() => {
    const element = elementFactory();
    element.setSelectionRange(1, 3, "backward");

    assert_equals(element.selectionStart, 1, "Sanity check: selectionStart was set correctly");
    assert_equals(element.selectionEnd, 3, "Sanity check: selectionEnd was set correctly");
    assert_equals(element.selectionDirection, "backward", "Sanity check: selectionDirection was set correctly");

    element.value = "hello";

    assert_equals(element.selectionStart, 1, "selectionStart must not change");
    assert_equals(element.selectionEnd, 3, "selectionEnd must not change");
    assert_equals(element.selectionDirection, "backward", "selectionDirection must not change");
  }, `${descriptor}: selection must not change when setting the same value`);

  test(() => {
    const element = elementFactory();
    element.setSelectionRange(1, 3, "backward");

    assert_equals(element.selectionStart, 1, "Sanity check: selectionStart was set correctly");
    assert_equals(element.selectionEnd, 3, "Sanity check: selectionEnd was set correctly");
    assert_equals(element.selectionDirection, "backward", "Sanity check: selectionDirection was set correctly");

    element.value = "hello2";

    assert_equals(element.selectionStart, element.value.length, "selectionStart must be reset to the end");
    assert_equals(element.selectionEnd, element.value.length, "selectionEnd must be reset to the end");
    assert_true(element.selectionDirection === "none" || element.selectionDirection === "forward",
      "selectionDirection must be set to either none or forward");
  }, `${descriptor}: selection must change when setting a different value`);
}

test(() => {
  const textarea = document.querySelector("#t1");
  textarea.value = "hell\no";
  textarea.setSelectionRange(1, 3, "backward");

  assert_equals(textarea.selectionStart, 1, "Sanity check: selectionStart was set correctly");
  assert_equals(textarea.selectionEnd, 3, "Sanity check: selectionEnd was set correctly");
  assert_equals(textarea.selectionDirection, "backward", "Sanity check: selectionDirection was set correctly");

  textarea.value = "hell\r\no";

  assert_equals(textarea.selectionStart, 1, "selectionStart must not change when setting to CRLF");
  assert_equals(textarea.selectionEnd, 3, "selectionEnd must not change when setting to CRLF");
  assert_equals(textarea.selectionDirection, "backward", "selectionDirection must not change when setting to CRLF");

  textarea.value = "hell\ro";

  assert_equals(textarea.selectionStart, 1, "selectionStart must not change when setting to CR");
  assert_equals(textarea.selectionEnd, 3, "selectionEnd must not change when setting to CR");
  assert_equals(textarea.selectionDirection, "backward", "selectionDirection must not change when setting to CR");
}, "textarea in document: selection must not change when setting the same normalized value");

</script>
