third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/textfieldselection/selection-after-content-change-expected.txt

This is a testharness.js-based test.
FAIL input out of document: selection must not change when setting the same value assert_equals: selectionStart must not change expected 1 but got 5
PASS input out of document: selection must change when setting a different value 
FAIL input in document: selection must not change when setting the same value assert_equals: selectionStart must not change expected 1 but got 5
PASS input in document: selection must change when setting a different value 
FAIL input in document, with focus: selection must not change when setting the same value assert_equals: selectionStart must not change expected 1 but got 5
PASS input in document, with focus: selection must change when setting a different value 
PASS textarea out of document: selection must not change when setting the same value 
PASS textarea out of document: selection must change when setting a different value 
PASS textarea in document: selection must not change when setting the same value 
PASS textarea in document: selection must change when setting a different value 
PASS textarea in document, with focus: selection must not change when setting the same value 
PASS textarea in document, with focus: selection must change when setting a different value 
PASS textarea in document: selection must not change when setting the same normalized value 
Harness: the test ran to completion.
